+++ docs/_ext/pytestdocs.py
@@ -1,0 +1,6 @@
+def setup(app):
+    app.add_crossref_type(
+        directivename="fixture",
+        rolename="fixture",
+        indextemplate="pair: %s; fixture",
+    )