+++ pytest_django_test/settings_sqlite.py
@@ -3,6 +3,6 @@ from .settings_base import *  # noqa
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': ':memory:'
+        'NAME': '/should_not_be_accessed',
     },
 }